Change 'getVariableFloat' to 'getVariableDouble' (#58)

+++ core-api/src/main/java/com/optimizely/ab/config/LiveVariable.java
@@ -69,8 +69,8 @@ public class LiveVariable implements IdKeyMapped {
         @SerializedName("string")
         STRING ("string"),
 
-        @SerializedName("float")
-        FLOAT ("float");
+        @SerializedName("double")
+        DOUBLE ("double");
 
         private final String variableType;
